Pouszeregowanie plikow i dodanie printow
