+++ Components/constans.py
@@ -1,0 +1,24 @@
+# === GPIO Pin Definitions ===
+TRIG_PIN = 10        # Trigger pin for ultrasonic sensor
+ECHO_PIN = 11        # Echo pin for ultrasonic sensor
+M1A = 26             # Motor 1 pin A
+M1B = 19             # Motor 1 pin B
+M2A = 13             # Motor 2 pin A
+M2B = 6              # Motor 2 pin B
+MZ1 = 7              # Motor Z pin A
+MZ2 = 8              # Motor Z pin B
+ENDSTOP1_PIN = 1     # Endstop start position 
+ENDSTOP2_PIN = 2     # Endstop end position 
+ENDSTOP3_PIN_MIN = 3     # Endstop Z homming
+ENDSTOP4_PIN_MAKS = 9
+SERVO_PIN = 20
+ENKODER_PIN_1 = 4
+ENKODER_PIN_2 = 5
+
+SLOW_THRESH = 0.5
+MID_THRESH = 1.0
+FAST_THRESH = 2.0
+
+SPEED_SLOW = 10
+SPEED_MID = 20
+SPEED_FAST = 30
