+++ main.py
@@ -1,0 +1,44 @@
+# This program demonstrates how to control two DC motors using the MDD10 HAT in Lock-Antiphase mode.
+# It uses an ultrasonic sensor to maintain a specified distance from a wall.
+# The robot moves forward or backward while adjusting its speed based on proportional-integral control (PI).
+
+import RPi.GPIO as GPIO          # Import GPIO library for Raspberry Pi
+from time import sleep, time     # Import sleep for delays and time for timestamps
+from Components import Robot
+GPIO.setmode(GPIO.BCM)           # Use Broadcom (BCM) pin numbering
+GPIO.setwarnings(False)          # Suppress GPIO warnings
+
+
+
+# Main setup
+
+robot = Robot()
+
+# === Main Test Sequence ===
+
+
+try:
+    while True:
+        print("Przód")
+        robot.Motor_Left.set_speed_motor(True, 70)
+        robot.Motor_Right.set_speed_motor(True, 70)
+        sleep(2)
+
+        print("Stop")
+        robot.Motor_Left.set_speed_motor(True, 0)
+        robot.Motor_Right.set_speed_motor(True, 0)
+        sleep(1)
+
+        print("Tył")
+        robot.Motor_Left.set_speed_motor(False, 70)
+        robot.Motor_Right.set_speed_motor(False, 70)
+        sleep(2)
+
+        print("Stop")
+        robot.Motor_Left.set_speed_motor(False, 0)
+        robot.Motor_Right.set_speed_motor(False, 0)
+        sleep(3)
+
+except KeyboardInterrupt:
+    GPIO.cleanup()
+    print("Program zatrzymany.")
